refactor(test) : testNumberGenerator 중복 코드 분리
