+++ src/test/java/racingcar/domain/CarTest.java
@@ -43,18 +43,4 @@ public class CarTest {
         //Then
         assertThat(car.getPosition()).isEqualTo(expectedResult);
     }
-
-    static class TestNumberGenerator implements NumberGenerator {
-
-        private final List<Integer> numbers;
-
-        public TestNumberGenerator(List<Integer> numbers) {
-            this.numbers = numbers;
-        }
-
-        @Override
-        public int generate() {
-            return numbers.remove(0);
-        }
-    }
 }
